fix type

+++ traffic.go
@@ -22,7 +22,7 @@ type BaremetalTraficHost struct {
 }
 
 // GetBaremetalHostTrafic ..
-func (api *API) GetBaremetalHostTrafic(id uint) (*BaremetalTraficHost, error) {
+func (api *API) GetBaremetalHostTrafic(id uint64) (*BaremetalTraficHost, error) {
 	var traffic BaremetalTraficHost
 	if err := api.getRequest(fmt.Sprintf("/hosts/%d/traffic_summary", id), &traffic); err != nil {
 		return nil, err
